modularized get id
